Reduce the warnings from Cygwin. Escpecially the change of 'struct in6_addr':   struct in6_addr {          union { -           u_char  _S6_u8[16]; -           u_short _S6_u16[8]; -           u_long  _S6_u32[4];  << !! this is 64-bit on LP64 +           u_int8_t  _S6_u8 [16]; +           u_int16_t _S6_u16 [8]; +           u_int32_t _S6_u32 [4];          } _S6_un;        };
was important for Cygwin-x64 since it's a LP64 platform.
Where 'sizeof(u_long) != sizeof(u_int)'.
Ref: https://en.wikipedia.org/wiki/64-bit_computing#64-bit_data_models

+++ src/ip4_frag.c
@@ -271,7 +271,7 @@ static __inline BOOL check_frag_ofs (const in_Header *ip, DWORD ofs, DWORD end)
      return (TRUE);
 
   TRACE_CONSOLE (2, "Bad frag-ofs: %lu, frag-end: %lu, ip-prot %u (%s -> %s)\n",
-                 ofs, end, ip->proto,
+                 (u_long)ofs, (u_long)end, ip->proto,
                  _inet_ntoa(NULL,intel(ip->source)),
                  _inet_ntoa(NULL,intel(ip->destination)));
   ARGSUSED (ip);
